PMM-13315 Move SanitizeSAName to global utils.
