+++ utils/strings/generate.go
